requirements.txt
accelerate==1.6.0
addict==2.4.0
airportsdata==20250224
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
appnope==0.1.4
argcomplete==3.6.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
bidict==0.23.1
bleach==6.2.0
blinker==1.9.0
blis==1.2.0
boto3==1.37.24
botocore==1.37.24
build==1.2.2.post1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
cohere==5.14.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
csvw==3.5.1
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.11
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
diskcache==5.6.3
distro==1.9.0
dlinfo==2.0.0
docopt==0.6.2
durationpy==0.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
espeakng-loader==0.2.4
eval_type_backport==0.2.2
executing==2.2.0
fastapi==0.115.11
fastavro==1.10.0
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.1.0
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fqdn==1.5.1
fsspec==2025.3.2
genson==1.3.0
gguf==0.16.2
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc0
google-api-python-client==2.168.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.69.1
griffe==1.7.1
groq==0.20.0
grpcio==1.71.0
grpcio-status==1.71.0
guidance-stitch==0.1.4
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
iniconfig==2.0.0
interegular==0.3.3
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
iso3166==2.1.1
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.14
kokoro==0.9.2
kubernetes==32.0.1
langcodes==3.5.0
language-tags==1.2.0
language_data==1.3.0
lark==1.2.2
llama_cpp_python @ file:///Users/ian/src/llama-cpp-python
llguidance==0.6.31
logfire-api==3.12.0
loguru==0.7.3
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mcp==1.6.0
mdurl==0.1.2
misaki==0.9.3
mistralai==1.6.0
mistune==3.1.3
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
murmurhash==1.0.12
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook_shim==0.2.4
num2words==0.5.14
numpy==2.2.5
oauthlib==3.2.2
ollama==0.4.7
onnxruntime==1.21.0
openai==1.70.0
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
ordered-set==4.1.0
orjson==3.10.15
outcome==1.3.0.post0
outlines==0.2.3
outlines_core==0.1.26
overrides==7.7.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
phonemizer-fork==3.3.2
pillow==11.1.0
pipdeptree==2.26.1
platformdirs==4.3.7
pluggy==1.5.0
posthog==3.19.0
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.4
pydantic-ai==0.0.48
pydantic-ai-slim==0.0.48
pydantic-evals==0.0.48
pydantic-graph==0.0.48
pydantic-settings==2.8.1
pydantic_core==2.33.2
pydot==3.0.4
pyformlang==1.0.11
Pygments==2.19.1
pyparsing==3.2.3
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-timeout==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-json-logger==3.3.0
python-socketio==5.12.1
PyYAML==6.0.2
pyzmq==26.3.0
rdflib==7.1.3
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.24.0
rsa==4.9
s3transfer==0.11.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
-e git+https://github.com/imars/seclorum.git@ed64f7953399342c561767c5d55f299c88daca5d#egg=seclorum
segments==2.3.0
selenium==4.29.0
Send2Trash==1.8.3
sentence-transformers==3.0.0
sentencepiece==0.2.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.6
spacy==3.8.4
spacy-curated-transformers==0.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.46.1
sympy==1.14.0
tenacity==9.0.0
terminado==0.18.1
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.9.0
timeout-decorator==0.5.0
tinycss2==1.4.0
tokenizers==0.21.1
torch==2.2.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
trio==0.29.0
trio-websocket==0.12.2
typer==0.15.2
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.0
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.0.4
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
wsproto==1.2.0
zipp==3.21.0
